+++ protocols/upnp/CHANGELOG.md
@@ -7,6 +7,8 @@
   with per-request in-flight tracking under a single `listener_id` key.
   See [PR 6459](https://github.com/libp2p/rust-libp2p/pull/6459)
 
+- Add integration tests. See [PR 6264](https://github.com/libp2p/rust-libp2p/pull/6264)
+
 ## 0.6.0
 
 - Change `Event::NewExternalAddr` and `Event::ExpiredExternalAddr` from tuple variants to struct variants